backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
eth-utils
eth-hash
aiofiles
spoon-ai-sdk
pydantic
pytest
